requirements.txt
chumpy
numpy==1.23.0
lightning==2.0.0
omegaconf==2.3.0
jaxtyping
typeguard
diffusers==0.21.0
transformers==4.26.1
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
libigl
xatlas
trimesh[easy]
networkx
pysdf # ensure libeigen3-dev is installled. Ubuntu: sudo apt-get install libeigen3-dev
PyMCubes==0.1.4
wandb
gradio==4.44.1
torchmetrics
pymeshlab==2022.2.post2
smplx==0.1.28
scikit-learn==1.3.2
yacs==0.1.8
open3d==0.18.0
iopath
fvcore

# deepfloyd
xformers==0.0.17
bitsandbytes==0.38.1
sentencepiece
safetensors
huggingface_hub==0.25.2

# for zero123
einops
kornia
taming-transformers-rom1504

#controlnet
controlnet_aux==0.0.7